+++ backend-lib/common-securities/src/main/java/com/optimize/common/securities/dto/UserDeviceRestrictionDto.java
@@ -1,0 +1,8 @@
+package com.optimize.common.securities.dto;
+
+import lombok.Data;
+
+@Data
+public class UserDeviceRestrictionDto {
+    private boolean enabled;
+}